move "lib/options" to root "settings" folder
